% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_conc_for_tissue_mass.R
\name{adjust_conc_for_tissue_mass}
\alias{adjust_conc_for_tissue_mass}
\title{Adjust compound concentrations for extracted tissue mass}
\usage{
adjust_conc_for_tissue_mass(
  data,
  tiss_mass,
  extract_vol = 1.5,
  prop_deriv = (1/1.5)
)
}
\arguments{
\item{data}{A data frame containing compound concentrations.}

\item{tiss_mass}{A vector of tissue mass extracted for each sample in data.}

\item{extract_vol}{Total volume of lipid extract.}

\item{prop_deriv}{Proportion of total lipid extract derivatized.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A data frame of tissue compound concentrations.
}
\description{
'adjust_conc_for_tissue_mass()' accepts compound concentration data from
chromatography and calculates the concentration of each compound in
extracted tissue.
}
\examples{
data <- data.frame(
  runif(5, min = 1, max = 10),
  runif(5, min = 1, max = 10),
  runif(5, min = 1, max = 10)
)
names(data) <- c("8:0", "10:0", "12:0")
tiss_mass <- rep(10, 5)

# Adjust concentration for tissue mass
adjust_conc_for_tissue_mass(data, tiss)

}
