% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gc_response_factor.R
\name{calc_gc_response_factor}
\alias{calc_gc_response_factor}
\title{Calculate gas chromatograph response factor}
\usage{
calc_gc_response_factor(data, ext_std_concs, ext_std_contents)
}
\arguments{
\item{data}{A data frame containing integrated chromatogram peak areas for
each fatty acid in external standards. Colnames must be fatty acid names.}

\item{ext_std_concs}{A numeric vector containing the concentrations of
external standards in ng/uL.}

\item{ext_std_contents}{A data frame containing fatty acids in external
standards and the proportion (decimal percentage) each fatty acid makes up
of standards. Proportions must sum to 1.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A df.
}
\description{
'calc_gc_response_factor()' calculates instrument response factors using
integrated chromatogram peak areas and external standard concentrations.
}
\examples{
# Provide some examples.
}
