% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_area_to_conc.R
\name{convert_area_to_conc}
\alias{convert_area_to_conc}
\title{Convert chromatogram peak area into concentration}
\usage{
convert_area_to_conc(data, rf_table, ..., rf_map = NULL)
}
\arguments{
\item{data}{A data frame containing peak areas.}

\item{rf_table}{A data frame containing instrument response factors.}

\item{...}{Additional arguments passed to methods.}

\item{rf_map}{A data frame containing response factor mappings.}
}
\value{
A data frame containing peak concentrations.
}
\description{
'convert_area_to_conc' converts chromatogram peak areas into concentrations
using an instrument response factor (RF) table for the external standards
run during batch processing. The function can also map RF values for fatty
acids (FA) not in external standards by passing in an RF mapping data
frame.
}
\examples{
# example compound peak area data
data <- data.frame(
 runif(5, min = 1000, max = 3000000),
 runif(5, min = 1000, max = 3000000),
 runif(5, min = 1000, max = 3000000),
 runif(5, min = 1000, max = 3000000),
 runif(5, min = 1000, max = 3000000),
 runif(5, min = 1000, max = 3000000)
)
names(data) <- c("8:0", "10:0", "12:0", "14:0", "i-15:0", "15:0")

# example response factor table
rf_table <- data.frame(
 fa = c("10:0", "8:0", "12:0", "14:0", "i-15:0", "15:0"),
 rf = c(15000, 15000, 16000, 12000, 0, 10000)
)

# example response factor map
rf_map <- data.frame(
 fa = c("8:0", "10:0", "12:0", "14:0", "i-15:0", "15:0"),
 ref_fa = c(NA, NA, NA, NA, "15:0", NA)
)

# Convert areas to concentration
convert_area_to_conc(data, rf_table, rf_map = rf_map)

}
