% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_result_to_prop.R
\name{convert_result_to_prop}
\alias{convert_result_to_prop}
\title{Convert chromatogram peak area or concentration into proportion}
\usage{
convert_result_to_prop(data, na.rm = TRUE)
}
\arguments{
\item{data}{A data frame containing numberic data}

\item{na.rm}{Logical. If TRUE, excludes NA or NaN values from calculation.}

\item{...}{Other arguments to be passed to or from methods.}
}
\value{
A data frame containing compound proportions.
}
\description{
'convert_result_to_prop()' converts chromatogram compound areas and
concentrations into proportions (decimal \% of all compounds for each
sample).
}
\examples{
# example data
data <- data.frame(
  runif(5, min = 1000, max = 300000),
  runif(5, min = 1000, max = 300000),
  runif(5, min = 1000, max = 300000),
  runif(5, min = 1000, max = 300000),
  runif(5, min = 1000, max = 300000)
)
names(data) <- c("8:0", "10:0", "12:0", "13:0", "14:0")

# Convert example data to proportions
convert_result_to_prop(data)

}
