% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_plot.R
\name{corr_plot}
\alias{corr_plot}
\title{Simple correlation plot with sensable defaults}
\usage{
corr_plot(corr, colorbreaks = c(-0.2, 0, 0.2, 0.4, 0.6, 0.8, 1),
  colors = NA, columnlabels = NA, labRow = NA, xlab = NA)
}
\arguments{
\item{corr}{correlation matrix}

\item{colorbreaks}{cutpoints of the discontinuous factor loading colors}

\item{colors}{a prespecified vector of colors corresponding to colorbreaks.}

\item{columnlabels}{x-axis labels (typically more discriptive variable/factor names)}

\item{labRow}{Names of variables / factors}

\item{xlab}{x-axis label}
}
\value{
A graphic of correlations
}
\description{
creates a simple correlation plot
    Essentially a heatmap built on top of gplot::heatmap.2. UNDER HEAVY DEVELOPMENT.
}
\examples{
library(datasets)
library(dplyr)
library(RColorBrewer)
data(mtcars)
mtcars[1:4,1:4] \%>\%
cor() \%>\%
corr_plot(columnlabels = 1:4)

}
