% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_cor_plot.R
\name{factor_cor_plot}
\alias{factor_cor_plot}
\title{Effective Factor Correlation Plot}
\usage{
factor_cor_plot(psych_object, colors = 5, names = NA, cl.lim = c(-1,
  1), tl.srt = 45, type = "lower")
}
\arguments{
\item{psych_object}{object resulting from psych::fa()}

\item{colors}{The number of colors to be utilized}

\item{names}{alternative factor names (vector)}

\item{cl.lim}{Vector range of corrleations to be plotted. Use c(0,1)}

\item{tl.srt}{Tilt of upper labels}

\item{type}{argument passed to corrplot}
}
\value{
A graphic of correlations
}
\description{
Plots inter-factor correlations from a psych object
}
\examples{
library(datasets)
corr.matrix <- cor(mtcars)
results <- psych::fa(corr.matrix, 5, rotate = "varimax")
factor_cor_plot(results, colors = 10)

}
