% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAdata.R
\name{FAdata}
\alias{FAdata}
\title{Function1: FAdata}
\usage{
FAdata(dt, convert = TRUE)
}
\arguments{
\item{dt}{data frame with ID (column 1), Species (column 2), and FA composition (column 3~n)}

\item{convert}{Please put "FALSE" here if you don't convert FA composition data to "sqrt"}
}
\description{
\code{FAdata} Make data frame for multivariate analysis; Make ID and Species vectors
}
