% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadings_table.R
\name{loadings_table}
\alias{loadings_table}
\title{Factor loadings table}
\usage{
loadings_table(loading_frame, loadings_no = 7, cutoff = 0.2, roundto = 3,
  trim = 0, Name = NA, Communalities = F)
}
\arguments{
\item{loading_frame}{The data frame to be edited}

\item{loadings_no}{the names of the row name.}

\item{cutoff}{loadings whose absolute value are less than this are excluded}

\item{roundto}{rounding to how many digits}

\item{trim}{Removes the first n characters of the Description column when rendering the table.}

\item{Name}{variable name (to be included as a column in data_dic)}

\item{Communalities}{Should the communalities be displayed?}
}
\value{
A table with rounded factor loadings, ommiting weak loadings with variable information on the side.
}
\description{
Creates a factor loading table
}
\note{
No longer exported, replaced by loadings_table_psych
}
