% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_FAdata.R
\name{plot_FAdata}
\alias{plot_FAdata}
\title{Function2: plot_FAdata}
\usage{
plot_FAdata(
  conv.data,
  method,
  print_stress = FALSE,
  print_fit = FALSE,
  plot_vector = FALSE,
  plot_legend = FALSE
)
}
\arguments{
\item{conv.data}{data frame made by "FAdata" function}

\item{method}{Please select dissimilarity index, such as "bray", "euclidean", "chao", "morisita" ...}

\item{print_stress}{Please put "TRUE" here if you see results of stress value}

\item{print_fit}{Please put "TRUE" here if you see table of vector analysis}

\item{plot_vector}{Please put "TRUE" here if you print vectors on nMDS plot}

\item{plot_legend}{Please put "TRUE" here if you print legend on nMDS plot}
}
\description{
\code{plot_FAdata} Make nMDS plot
}
