% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scree_plot.R
\name{scree_plot}
\alias{scree_plot}
\title{Scree plot}
\usage{
scree_plot(corr, observations = nrow(corr), variables = ncol(corr))
}
\arguments{
\item{corr}{The correlation matrix to be used}

\item{observations}{number of observations in dataset}

\item{variables}{number of variables in dataset}
}
\value{
A plottable scree plot object
}
\description{
Easy, one function approach to creating scree plots
}
\examples{
set.seed(100)
data <- matrix(rnorm(100), ncol = 10) # loading our 'data'
data_cor <- cor(data)
scree <- scree_plot(data_cor, 10, 10)
plot(scree)

}
