% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfun.R
\name{testfun}
\alias{testfun}
\title{Add some numbers}
\usage{
testfun(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{Another number.}
}
\value{
The sum.
}
\description{
Add some numbers
}
\examples{
x <- 5
y <- 6
testfun(x, y) # returns 11
}
