% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gc_response_factor.R
\name{calc_gc_response_factor}
\alias{calc_gc_response_factor}
\title{Calculate gas chromatograph response factor}
\usage{
calc_gc_response_factor(data, ext_std_concs, ext_std_contents)
}
\arguments{
\item{data}{A data frame containing integrated chromatogram peak areas for
each fatty acid in external standards. Colnames must be fatty acid names.}

\item{ext_std_concs}{A numeric vector containing the concentrations of
external standards in ng/uL.}

\item{ext_std_contents}{A data frame containing fatty acids in external
standards and the proportion (decimal percentage) each fatty acid makes up
of standards. Proportions must sum to 1.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A data frame of fatty acids and their calculated response factors.
}
\description{
'calc_gc_response_factor()' calculates instrument response factors using
integrated chromatogram peak areas and external standard concentrations.
}
\examples{
# Make example peak area results
peak_areas <- data.frame(
  c(59690, 197032, 394075, 1041447),
  c(124260, 410732, 851543, 2092255),
  c(249962, 869479, 1895680, 4448913)
)

colnames(
  peak_areas
) <- c("8.0", "10.0", "16.0")

# Make example external standards proportions list
ext_standard_proportions <- data.frame(
  fa = c("8.0", "10.0", "16.0"),
  prop = as.numeric(c(0.01, 0.02, 0.05))
)

# Calculate response factors
calc_gc_response_factor(
  peak_areas,
  ext_std_concs = c(15, 50, 100, 250),
  ext_std_contents = ext_standard_proportions
)

}
