% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_area_to_conc.R
\name{convert_area_to_conc}
\alias{convert_area_to_conc}
\title{Convert chromatogram peak area into concentration}
\usage{
convert_area_to_conc(data, rf_table, ..., rf_map = NULL)
}
\arguments{
\item{data}{A data frame containing peak areas.}

\item{rf_table}{A data frame containing instrument response factors.}

\item{...}{Additional arguments passed to methods.}

\item{rf_map}{A data frame containing response factor mappings.}
}
\value{
A data frame containing peak concentrations.
}
\description{
'convert_area_to_conc' converts chromatogram peak areas into concentrations
using an instrument response factor (RF) table for the external standards
run during batch processing. The function can also map RF values for fatty
acids (FA) not in external standards by passing in an RF mapping data
frame.
}
\examples{
# Add some examples

}
