% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_fa_name.R
\name{convert_fa_name}
\alias{convert_fa_name}
\title{Convert fatty acid names into common formats}
\usage{
convert_fa_name(x, style = 1, notation = "ω", sep = ":", prefsep = "-")
}
\arguments{
\item{x}{A vector of fatty acid names.}

\item{style}{A number corresponding to a style table that dictates how names
will be converted. Defaults to "1".}

\item{notation}{A character vector that determines what precedes the
numerical representation of the location of double bonds.}

\item{sep}{A character vector that determines what separator is used
between carbon length and number of desaturations.}

\item{prefsep}{A character vector that determines what separator is used
after iso and anteiso prefixes, if applicable.}
}
\value{
A vector of reformatted names.
}
\description{
'convert_fa_name()' converts fatty acid names into commonly used formats and
provides the ability to easily customize how they are written.
}
\examples{
x <- c("c16.1w7c", "18.0", "20_1_w9", "i 15:0")

# returns c("16:1ω7c", "18:0", "20:1ω9", "i-15:0")
convert_fa_name(x)

# returns c("16.1 (n-7) c", "18.0", "20.1 (n-9)", "i-15.0")
convert_fa_name(x, style = 3, notation = "n", sep = ".")
}
