% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_fa_name.R
\name{find_fa_name}
\alias{find_fa_name}
\title{Find index of fatty acid name}
\usage{
find_fa_name(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
An integer vector of indexes where fatty acid names are found in x.
}
\description{
'find_fa_name()' finds fatty acid names using regular
expressions and returns a integer vector of
indexes where fatty acid names were found.
}
\examples{
# Character vector containing some fatty acid names
x <- c("c16.1w7c", "not-a-fa", "sample_id", "18.0", "20_1_w9", "i 15:0")

# Returns integer vector c(1L, 4L, 5L, 6L)
find_fa_name(x)
}
