% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadings_plot.R
\name{loadings_plot}
\alias{loadings_plot}
\title{Factor loadings plot}
\usage{
loadings_plot(loadings, columnlabels, colorbreaks = c(0, 0.2, 0.4, 0.6, 0.8,
  1), colors = brewer.pal(length(colorbreaks), "Greens"))
}
\arguments{
\item{loadings}{factor loadings}

\item{columnlabels}{x-axis labels (typically more discriptive variable names)}

\item{colorbreaks}{cutpoints of the discontinuous factor loading colors}

\item{colors}{colors pertaining to the 'colorbreaks' argument}
}
\value{
A graphic of factor loadings.
}
\description{
Creates a factor loading plot.
Essentially a heatmap built on top of gplot::heatmap.2.
}

