% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadings_plot.R
\name{loadings_plot}
\alias{loadings_plot}
\title{Factor loadings plot}
\usage{
loadings_plot(loadings, colorbreaks = c(0, 0.2, 0.4, 0.6, 0.8, 1),
  colors = NA, columnlabels = NA, labRow = NA, xlab = NA)
}
\arguments{
\item{loadings}{factor loadings}

\item{colorbreaks}{cutpoints of the discontinuous factor loading colors}

\item{colors}{a prespecified vector of colors corresponding to colorbreaks.}

\item{columnlabels}{x-axis labels (typically more discriptive variable names)}

\item{labRow}{Names of factors (defaults to factor 1, factor 2, etc.)}

\item{xlab}{x-axis label}
}
\value{
A graphic of factor loadings.
}
\description{
Creates a factor loading plot.
    Essentially a heatmap built on top of gplot::heatmap.2. UNDER HEAVY DEVELOPMENT.
}
\examples{
library(datasets)
corr.matrix <- cor(mtcars)
results <- psych::fa(corr.matrix, 5, rotate = "varimax")
loadings_plot(loadings = results$loadings,
              colorbreaks = c(0, 0.2, 0.4, 0.6, 0.8,1),
              colors = NA,
              columnlabels = colnames(mtcars),
              labRow = NA,
              xlab = "Variables")

}
