% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadings_table.R
\name{loadings_table}
\alias{loadings_table}
\title{Factor loadings table}
\usage{
loadings_table(loading_frame, loadings_no = ncol(loading_frame),
  cutoff = 0.2, roundto = 3, trim = 0, Name = NA,
  Communalities = F, description = NA)
}
\arguments{
\item{loading_frame}{The data frame to be edited}

\item{loadings_no}{the names of the row name.}

\item{cutoff}{loadings whose absolute value are less than this are excluded}

\item{roundto}{rounding to how many digits}

\item{trim}{Removes the first n characters of the Description column when rendering the table.}

\item{Name}{variable name (to be included as a column in data_dic)}

\item{Communalities}{Should the communalities be displayed?}

\item{description}{variable description (optional)}
}
\value{
A table with rounded factor loadings, ommiting weak loadings with variable information on the side.
}
\description{
Creates a factor loading table
}
\note{
possibly will be replaced by loadings_table_psych
}
