% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadings_table_psych.R
\name{loadings_table_psych}
\alias{loadings_table_psych}
\title{Factor loadings table from psych object}
\usage{
loadings_table_psych(psych_object, cutoff = 0.2, roundto = 3, trim = 0,
  fac.lab = NULL, var.lab = NULL, communalities = T)
}
\arguments{
\item{psych_object}{The Factor analysis psych object}

\item{cutoff}{loadings whose absolute value are less than this are excluded}

\item{roundto}{rounding to how many digits}

\item{trim}{Removes the first n characters of the Description column when rendering the table.}

\item{fac.lab}{Alternative factor labels}

\item{var.lab}{Alternative variable labels}

\item{communalities}{include communalities?}
}
\value{
A table with rounded factor loadings, ommiting weak loadings with variable information on the side.
}
\description{
Creates a factor loading table
}
\examples{
library(datasets)
corr.matrix <- cor(mtcars)
results <- psych::fa(corr.matrix, 3, rotate = "varimax")
loadings_table_psych(results, cutoff = 0.3, roundto = 2, fac.lab = c("F1", "F2", "F3"))

}
