% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatools.R
\name{scree_plot}
\alias{scree_plot}
\title{Scree plot}
\usage{
scree_plot(corr, observations, variables)
}
\arguments{
\item{corr}{The correlation matrix to be used}

\item{observations}{number of observations in dataset}

\item{variables}{number of variables in dataset}
}
\value{
a plot
}
\description{
Easy, one function approach to creating scree plots
}
\examples{
df_rename(
     data.frame(letters[1:5], 1:5),
     LETTERS[1:5],
     "CAPS"
)

}

